Move tests to individual fixtures - flip-comparisons
