+++ packages/babel-plugin-minify-flip-comparisons/__tests__/fixtures/binary-expr-constants-first/actual.js
@@ -1,0 +1,9 @@
+// constants first
+a === -1;
+
+x * 100;
+x + 100;
+x - 100;
+x / 100;
+x > 100;
+x === void 0;
