+++ packages/babel-plugin-minify-flip-comparisons/__tests__/fixtures/binary-expr-constants-first/expected.js
@@ -1,0 +1,9 @@
+// constants first
+-1 === a;
+
+100 * x;
+x + 100;
+x - 100;
+x / 100;
+100 < x;
+void 0 === x;
