+++ packages/babel-plugin-minify-flip-comparisons/__tests__/fixtures/binary-expr-pures-first/actual.js
@@ -1,0 +1,14 @@
+// pures first
+
+a === null;
+
+a === {};
+
+function foo() {
+  if (foo !== null) {
+    var bar;
+    bar = baz;
+  }
+  x();
+  return x;
+}
