+++ packages/babel-plugin-minify-flip-comparisons/__tests__/fixtures/binary-expr-pures-first/expected.js
@@ -1,0 +1,14 @@
+// pures first
+
+null === a;
+
+({}) === a;
+
+function foo() {
+  if (null !== foo) {
+    var bar;
+    bar = baz;
+  }
+  x();
+  return x;
+}
