+++ packages/babel-plugin-minify-flip-comparisons/__tests__/fixtures/binary-expr-values-first/expected.js
@@ -1,0 +1,2 @@
+// values first
+1 === a;
