+++ packages/babel-plugin-minify-flip-comparisons/__tests__/fixtures/conditionals-same-consequent/actual.js
@@ -1,0 +1,1 @@
+x === null ? undefined : x === undefined ? undefined : x ? foo(x) : wat();
