+++ packages/babel-plugin-minify-flip-comparisons/__tests__/fixtures/conditionals-same-consequent/expected.js
@@ -1,0 +1,1 @@
+null === x ? undefined : x === undefined ? undefined : x ? foo(x) : wat();
